Refactor: Move TripRequest model to dedicated schemas module
