+++ uvicorn.out
@@ -1,0 +1,5 @@
+INFO:     Started server process [8300]
+INFO:     Waiting for application startup.
+INFO:     Application startup complete.
+INFO:     Uvicorn running on http://0.0.0.0:8000 (Press CTRL+C to quit)
+INFO:     127.0.0.1:43896 - "GET / HTTP/1.1" 200 OK
